Faster chains of Richelot isogenies
The code was repeatedly taking the 2^(a-1)-th power
of a point on the Jacobian. But since isogenies are group morphisms,
the computation of these powers on each Jacobian is somehow redundant.

By precomputing a power, it can be faster to map it through the
isogenies rather than starting over from generators each time.

The code executes about 2x-3x faster for large key sizes
and can solve SIKEp434 in less than 30 minutes.