Showcase how to attack when 2^a-3^b = k(u^2+v^2) where k is small/smooth
Combining ideas of Castryck-Decru with Maino-Martindale,
we can see that when c=k(u^2+v^2) the attack is just as effective
if we know how to produce a degree k isogeny from E_start.

In the case of SIKEp610 we can arrange for k=7, which is an easy
precomputation, reducing the necessary guesses from 3^5 to 3^4.

This reduces the worst case duration from 25 minutes to about 8 minutes
on a single ordinary desktop core.